requirements.txt
# Core web framework
fastapi
uvicorn

# Data & validation
pydantic
python-multipart

# Database
sqlalchemy
databases

# Spatial / raster processing
shapely
rasterio
geopandas

# Optimization engine
pulp

# Utilities
requests
python-dotenv